requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.5.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
narwhals==2.5.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.17.0
packaging==23.2
pandas==2.3.2
Pillow==9.5.0
plotly==6.3.0
protobuf==4.25.8
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
referencing==0.36.2
requests==2.32.5
rich==13.9.4
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.5.0
validators==0.35.0
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
